requirements.txt
easyocr==1.6.2
numpy==1.23.3
Pillow==9.3.0
PyAutoGUI==0.9.54
pynput==1.7.6
pyperclip==1.8.2
pystray==0.19.4
screeninfo==0.8.1
torch==1.13.0
win10toast==0.9
googletrans==4.0.0rc1


# For desktop setup, due to package conflicts, you may need to run these command after install above packages.
# pip uninstall  opencv-python 
# pip uninstall opencv-python-headless 
# pip install opencv-python==4.6.0.66